rename: reward 관련 general dto 위치 변경
